+++ computer graphics/includes/Operators.h
@@ -5,21 +5,21 @@
 #include <iostream>
 
 
-double clamp(double val, double minVal, double maxVal)
+float clamp(float val, float minVal, float maxVal)
 {
 	using namespace std;
 	return max(minVal, min(maxVal, val));
 }
 
-double remap(double val, double inMin, double inMax, double outMin, double outMax)
+float remap(float val, float inMin, float inMax, float outMin, float outMax)
 {
 	return outMin + (outMax - outMin) * ((val - inMin) / (inMax - inMin));
 }
 
 template<typename t>
-t perspectiveCorrectInterpolateAttribute(t a0, t a1, t a2, double v0_w, double v1_w, double v2_w, double alpha, double beta, double gamma)
+t perspectiveCorrectInterpolateAttribute(t a0, t a1, t a2, float v0_w, float v1_w, float v2_w, float alpha, float beta, float gamma)
 {
-	double frag_w = ((alpha * v0_w) + (beta * v1_w) + (gamma * v2_w));
+	float frag_w = ((alpha * v0_w) + (beta * v1_w) + (gamma * v2_w));
 	t attrib[3];
 	attrib[0] = a0 * alpha * v0_w;
 	attrib[1] = a1 * beta * v1_w;
